restricoes de logged in no site inteiro, username nao podem ter espacos
imagem para a conta do runlo, nome de museus ja nao pode ser alterado

+++ reject.php
@@ -1,9 +1,19 @@
 <?php
 require_once("connectDB.php");
 
-$id = str_replace('-', ' ', $_POST["reqID"]);
+if(isSet($_SESSION['username'])){
+  if(isSet($_SESSION['loginErro']) && $_SESSION['loginErro'] == 0){
+    if($_SESSION['type'] == "admin"){
 
-mysqli_query($connection,"DELETE FROM requests WHERE id = '$id'");
-header('Location: http://localhost/LI4/requests.php');
+
+    $id = str_replace('-', ' ', $_POST["reqID"]);
+
+    mysqli_query($connection,"DELETE FROM requests WHERE id = '$id'");
+    header('Location: http://localhost/LI4/requests.php');
+    }else{
+      echo "Permition Denied";
+    }
+  }
+}else header('Location: http://localhost/LI4/login.php');
 
 ?>
